requirements.txt
asgiref==3.11.1
Django==4.2.28
django-cors-headers==4.9.0
djangorestframework==3.16.1
psycopg==3.2.13
psycopg-binary==3.2.13
python-dotenv==1.2.1
sqlparse==0.5.5
typing_extensions==4.15.0
